+++ web/src/user/settings/routes.tsx
@@ -57,11 +57,4 @@ export const userSettingsAreaRoutes: readonly UserSettingsAreaRoute[] = [
         ),
         condition: () => window.context.accessTokensAllow !== 'none',
     },
-    {
-        path: '/permissions',
-        exact: true,
-        render: lazyComponent(() => import('./auth/UserSettingsPermissionsPage'), 'UserSettingsPermissionsPage'),
-        condition: ({ authenticatedUser }) =>
-            !!window.context.site['permissions.backgroundSync']?.enabled && authenticatedUser.siteAdmin,
-    },
 ]
